Refactoring: xml parsing bound to CustomElement instead of SVG
